# v18 test runs/build-b-codex/contracts/openapi/previous.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Project Scaffolding + Prisma Schema API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "tags": [
          "api"
        ],
        "responses": {
          "201": {
            "description": "Generated response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "register",
        "tags": [
          "api"
        ],
        "responses": {
          "201": {
            "description": "Generated response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "name",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "me",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Generated response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "task_id": {
                      "type": "string"
                    },
                    "author_id": {
                      "type": "string"
                    },
                    "author": {
                      "$ref": "#/components/schemas/UserSummaryDto"
                    },
                    "created_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "content",
                    "task_id",
                    "author_id",
                    "author",
                    "created_at"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  }
                },
                "required": [
                  "content"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "operationId": "findAll",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "findOne",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/auth/login": {
      "post": {
        "operationId": "login",
        "tags": [
          "auth"
        ],
        "responses": {
          "201": {
            "description": "Generated response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register",
        "tags": [
          "auth"
        ],
        "responses": {
          "201": {
            "description": "Generated response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "name",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "me",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Generated response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "task_id": {
                      "type": "string"
                    },
                    "author_id": {
                      "type": "string"
                    },
                    "author": {
                      "$ref": "#/components/schemas/UserSummaryDto"
                    },
                    "created_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "content",
                    "task_id",
                    "author_id",
                    "author",
                    "created_at"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  }
                },
                "required": [
                  "content"
                ]
              }
            }
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "operationId": "findOne",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "Generated response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
...
}